Replaced arg of first Mapper by LongWritable

+++ code/lsimr/src/main/java/org/lsi/mapreduce/ConnectedComponentsCounter.java
@@ -12,6 +12,7 @@ import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.filecache.DistributedCache;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.io.Writable;
 import org.apache.hadoop.mapred.FileInputFormat;
@@ -33,7 +34,7 @@ import org.apache.hadoop.util.ToolRunner;
 public class ConnectedComponentsCounter extends Configured implements Tool {
 
 	public static class MapFirstPass extends MapReduceBase implements
-			Mapper<IntWritable, Text, IntWritable, IntIntWritableTuple> {
+			Mapper<LongWritable, Text, IntWritable, IntIntWritableTuple> {
 
 		private IntIntWritableTuple idAndParentCell = new IntIntWritableTuple();
 		private IntWritable idColumn = new IntWritable();
@@ -54,7 +55,7 @@ public class ConnectedComponentsCounter extends Configured implements Tool {
 
 		// <get byte offset in input line, text of a line>
 		// Return <idcolumn;<idcell,booleancell>>
-		public void map(IntWritable key, Text value,
+		public void map(LongWritable key, Text value,
 				OutputCollector<IntWritable, IntIntWritableTuple> output,
 				Reporter reporter) throws IOException {
